# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engine/states:
    get:
      operationId: getEngineStates
      tags: [engine]
      summary: Read per-symbol strategy states
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Strategy state list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngineState"
        "400":
          $ref: "#/components/responses/BadRequest"
  /engine/run:
    post:
      operationId: runEngine
      tags: [engine]
      summary: Analyze live candles and update strategy state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EngineRunInput"
      responses:
        "200":
          description: Updated strategy states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EngineState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /engine/backtest:
    post:
      operationId: runBacktest
      tags: [engine]
      summary: Replay real broker candles with explicit trading costs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "200":
          description: Historical replay result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /engine/execute:
    post:
      operationId: executeTrade
      tags: [engine]
      summary: Execute a protected live MetaApi order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeExecuteInput"
      responses:
        "200":
          description: Executed trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeExecution"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /journal:
    get:
      operationId: getJournal
      tags: [journal]
      summary: Read journal entries
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
  /equity-history:
    get:
      operationId: getEquityHistory
      tags: [journal]
      summary: Read equity history
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Equity history points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquityPoint"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    AccountId:
      name: accountId
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Trading gate blocked the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UpstreamError:
      description: Upstream service failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: [string, "null"]
      required: [error]
...
    EngineRunInput:
      type: object
      properties:
        accountId:
          type: string
          minLength: 1
        symbols:
          type: array
          items:
            type: string
          minItems: 1
      required: [accountId, symbols]
    BacktestInput:
      type: object
      properties:
        accountId:
          type: string
          minLength: 1
        symbol:
          type: string
          minLength: 1
        timeframe:
          type: string
          enum: [1m, 5m, 15m, 1h, 4h, 1d]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        startingBalance:
          type: number
          exclusiveMinimum: 0
        riskPerTrade:
          type: number
          exclusiveMinimum: 0
        spreadPoints:
          type: number
          minimum: 0
        slippagePoints:
          type: number
          minimum: 0
        commissionPerLot:
          type: number
          minimum: 0
      required: [accountId, symbol, timeframe, startTime, endTime, startingBalance, riskPerTrade, spreadPoints, slippagePoints, commissionPerLot]
    BacktestTrade:
      type: object
      properties:
        direction:
          type: string
          enum: [BUY, SELL]
        entryTime:
          type: string
        exitTime:
          type: [string, "null"]
        entry:
          type: number
        exit:
          type: [number, "null"]
        sl:
          type: number
        tp:
          type: number
        lot:
          type: number
        pnl:
          type: [number, "null"]
        outcome:
          type: string
          enum: [WIN, LOSS, OPEN]
      required: [direction, entryTime, entry, sl, tp, lot, outcome]
    BacktestResult:
      type: object
      properties:
        symbol:
          type: string
        realSymbol:
          type: string
        timeframe:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        trades:
          type: array
          items:
            $ref: "#/components/schemas/BacktestTrade"
        startingBalance:
          type: number
        endingBalance:
          type: number
        netPnl:
          type: number
        maxDrawdown:
          type: number
        winRate:
          type: [number, "null"]
        profitFactor:
          type: [number, "null"]
        candleCount:
          type: number
        dataSource:
          type: string
      required: [symbol, realSymbol, timeframe, startTime, endTime, trades, startingBalance, endingBalance, netPnl, maxDrawdown, candleCount, dataSource]
    EngineState:
      type: object
      properties:
        id:
          type: [string, "null"]
        accountId:
          type: string
        symbol:
          type: string
        currentState:
          type: string
        htfBias:
          type: [string, "null"]
        htfConflict:
          type: boolean
        trend:
          type: [string, "null"]
        poiType:
          type: [string, "null"]
        poiHigh:
          type: [number, "null"]
        poiLow:
          type: [number, "null"]
        diagnostics:
          type: array
          items:
            type: string
        lastUpdated:
          type: [string, "null"]
      required: [accountId, symbol, currentState, htfConflict, diagnostics]
    TradeExecuteInput:
      type: object
      properties:
        accountId:
          type: string
          minLength: 1
        symbol:
          type: string
        realSymbol:
          type: string
        direction:
          type: string
          enum: [BUY, SELL]
        lot:
          type: number
          exclusiveMinimum: 0
        sl:
          type: number
        tp:
          type: number
        poiType:
          type: [string, "null"]
        bosMssTag:
          type: [string, "null"]
      required: [accountId, symbol, realSymbol, direction, lot, sl, tp]
    TradeExecution:
      type: object
      properties:
        orderId:
          type: [string, "null"]
        positionId:
          type: [string, "null"]
        accountId:
          type: string
        symbol:
          type: string
        direction:
          type: string
        lot:
          type: number
        entry:
          type: [number, "null"]
        sl:
          type: number
        tp:
          type: number
        leverage:
          type: [number, "null"]
        marginUsed:
          type: [number, "null"]
        status:
          type: string
      required: [accountId, symbol, direction, lot, sl, tp, status]
    JournalEntry:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        symbol:
          type: string
        realSymbol:
          type: [string, "null"]
        direction:
          type: string
        entry:
          type: [number, "null"]
        sl:
          type: [number, "null"]
        tp:
          type: [number, "null"]
        lot:
          type: [number, "null"]
        pnl:
          type: [number, "null"]
        rMultiple:
          type: [number, "null"]
        status:
          type: string
        poiType:
          type: [string, "null"]
        bosMssTag:
          type: [string, "null"]
        htfBias:
          type: [string, "null"]
        leverage:
          type: [number, "null"]
        marginUsed:
          type: [number, "null"]
        chartSnapshotUrl:
          type: [string, "null"]
        createdAt:
          type: [string, "null"]
      required: [id, accountId, symbol, direction, status]
    EquityPoint:
      type: object
      properties:
        timestamp:
          type: string
        balance:
          type: number
        equity:
          type: number
      required: [timestamp, balance, equity]
...