# lib/api-zod/src/generated/types/engineRunInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EngineRunInput {
  /** @minLength 1 */
  accountId: string;
  /** @minItems 1 */
  symbols: string[];
}